\name{Internal functions and methods}
\alias{schlatherfull}
\alias{smithfull}
\alias{dpostmaxstab}
\alias{dpriormaxstab}
\alias{smithform}
\alias{schlatherform}
\alias{schlatherindform}
\alias{schlatherindfull}
\alias{geomgaussform}
\alias{geomgaussfull}
\alias{fitnsmaxstab}
\alias{nsgeomgaussfull}
\alias{nsgeomgaussform}
\alias{brownresnickfull}
\alias{brownresnickform}
\alias{extremaltform}
\alias{extremaltfull}
\alias{getsubregions}

\title{Internal functions and methods for the maxstable package.}
\description{A set of functions that should not be used directly by the
user or are not documented yet. For methods, user should usually used
the generic functions which calls the appropriate method.}
\author{Mathieu Ribatet}
\keyword{internal}
