% Generated by roxygen2: do not edit by hand
% Please edit documentation in ./<text>
\name{predict.spatgev}
\alias{predict.spatgev}
\title{Predict Method for the \code{"spatgev"} Class}
\usage{
\method{predict}{spatgev}(
  object,
  newdata,
  temp.newdata,
  keep = NULL,
  temp.keep = NULL,
  ret.per = c(10, 50, 100),
  level = 0.95,
  out = c("data", "array"),
  ...
)
}
\arguments{
\item{object}{An object with class \code{"spatgev"}.}

\item{newdata}{A matrix or data frame with the new "spatial"
covariates.}

\item{temp.newdata}{An optional matrix or data frame with the new
"temporal" covariates.}

\item{keep, temp.keep}{Optional character vectors giving the name
of columns to be kept in the output. These will actually be
used only when \code{out} specifies a data frame output
i.e. when \code{out} is \code{"data"}. \bold{Not implemented
yet}.}

\item{ret.per}{A vector of return periods.}

\item{level}{A confidence level between 0.0 and 1.0. If
\code{level} is \code{NULL} or has length zero, the confidence
limits will not be computed.}

\item{out}{Character telling what kind of output is wanted. See
\bold{Details}.}

\item{...}{Not used yet.}
}
\value{
A data frame or a list of arrays depending on the value of
    \code{out}, see \bold{Details}.
}
\description{
Predict method for the \code{"spatgev"} class.
}
\details{
The output depends on the choice given by the formal
    argument \code{out}. In both cases this is a list with two
    elements \code{GEVParam} and \code{RL} giving the GEV
    parameters and the Return Levels (provided that \code{ret.lev}
    was not of length zero).

\itemize{ 
    \item{\code{"data"} }{
        The elements \code{GEVParam} and \code{RL} are \emph{data
        frames}. Both have columns with understandable names such
        as \code{"RL_Est"} for the estimated return level
        a.k.a. quantile.  This choice is relevant when plots are
        to be produced using \pkg{ggplot2}. The variables
        specified with \code{keep} can then be used to define the
        aesthetics or the facets in trellis graphics.
     } 
     \item{\code{"array"} }{
        The elements \code{GEVParam} and \code{RL} are two
        \emph{arrays}. In both cases the first two dimensions
        match the spatial and the temporal dimensions. The other
        dimensions give the the return period, the return level,
        the type of return level (estimate, lower upper bound) or
        the GEV parameter (with value \code{"loc"} \code{"scale"}
        or \code{"shape"}).
    }
}
}
\section{Caution}{
 This method will work correctly only it at the
    creation of \code{object} the values 
    \itemize{
    \item{\code{data} }{
        Should have rownames allowing the identification of the
        observations. These can be years or dates in POSIX format
        like \code{"2020-01-01"}. Should also have colnames
        allowing the identification of the sites.
    }
    \item{\code{covariables} }{
        Should have rownames allowing the identification of the
        sites, identical to the colnames of \code{data}.
    }
   \item{\code{temp.data} } {
        If used whould have rownames allowing the identification of
        the observations, identical to the rownames of \code{data}.
   }
   }
   The same rules apply to \code{newdata} and \code{temp.newdata}.
   Note that when \code{temp.data} has been used at the creation of
   \code{object}, the formal argument \code{temp.newdata} should
   be used in \code{predict}.
}
 
